pyproject.toml
[tool.poetry]
name = "weight-tracker"
version = "0.1.0"
description = "Weight Tracker app"
authors = ["Justas Zieminykas <justaxsz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.167.0"
pandas = "^2.2.3"
myfitnesspal = "^2.1.2"
ipykernel = "^6.29.5"
flask-cors = "^6.0.1"
mypy = "^1.17.1"
google-auth-stubs = "^0.3.0"
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
pytest-cov = "^6.2.1"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"

[tool.mypy]
mypy_path = "app"
namespace_packages = true
explicit_package_bases = true
exclude = [
    "app/throwaway/.*\\.py$",
    "app/data/.*\\.py$",
    "app/static/.*\\.py$",
    "app/templates/.*\\.py$",
    "app/_pycache__/.*\\.py$",
    "venv/",
    ".venv/",
    "app/app.py",
]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pandas-stubs = "^2.3.0.250703"
google-api-python-client-stubs = "^1.30.0"
ruff = "^0.12.11"
black = "^25.1.0"

[tool.ruff]
line-length = 88
target-version = "py312"
src = ['app']

[tool.ruff.lint]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "UP",  # pyupgrade
  "SIM", # flake8-simplify
  "RUF", # Ruff-specific
]
exclude = [
    "./app/throwaway",
    "app/data",
    "app/static",
    "app/templates",
    "app/_pycache__",
    "app/app.py",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = ["app"]
python_files = ["test_*.py"]

[tool.coverage.run]
omit = [
  "app/app.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
